refactor(ui): keep query helpers form-local
