+++ ui/src/app/workflows/sitecablevalidationworkflow/form/site-cable-validation-workflow-form.tsx
@@ -19,7 +19,7 @@
 import * as React from "react";
 import { useSearchParams } from "next/navigation";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { useForm } from "react-hook-form";
+import { useForm, type UseFormReturn } from "react-hook-form";
 import { z } from "zod";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
@@ -29,10 +29,6 @@ import {
 import { useToast } from "@/components/ui/use-toast";
 import { SiteCableValidationWorkflowInput } from "@/types/data-table.types";
 import { useEnvData } from "@/hooks";
-import {
-  setMultiQueryValue,
-  setSingleQueryValue,
-} from "@/lib/query-form-helpers";
 import { getErrorMessage, startWorkflow } from "@/lib/utils";
 import { WorkflowFormField } from "@/components/forms/formfield";
 
@@ -48,6 +44,34 @@ const SiteCableValidationFormSchema = z.object({
 });
 
 type SiteCableValidationFormData = z.infer<typeof SiteCableValidationFormSchema>;
+type QueryOption = { key: string; value: string };
+type SingleValueField = "site" | "tenant";
+type MultiValueField = "roles" | "status";
+
+/** Maps a single query parameter key to its form option value or clears it. */
+const setSingleQueryValue = (
+  form: UseFormReturn<SiteCableValidationFormData>,
+  queryValue: string | null,
+  options: QueryOption[],
+  fieldName: SingleValueField
+) => {
+  const fieldValue =
+    options.find((option) => option.key === queryValue)?.value ?? "";
+  form.setValue(fieldName, fieldValue);
+};
+
+/** Keeps valid multi-value query keys and clears values without an option. */
+const setMultiQueryValue = (
+  form: UseFormReturn<SiteCableValidationFormData>,
+  queryValues: string[],
+  options: QueryOption[],
+  fieldName: MultiValueField
+) => {
+  const fieldValue = queryValues.filter((queryValue) =>
+    options.some((option) => option.key === queryValue)
+  );
+  form.setValue(fieldName, fieldValue);
+};
 
 /** Renders the form for starting a site cable validation workflow. */
 export const SiteCableValidationWorkflowForm = () => {
